requirements.txt
-r ddhldap/requirements.txt
-r wagtailbase/requirements.txt
Fabric
Pillow
Unidecode
beautifulsoup4
bleach
django==3.2
django-appconf>=1.0.3
django-author==1.0.1
django-extensions==1.9.7
django-grappelli==2.15.1
django-model-utils==4.0.0
django-mptt==0.13.3
django-mptt-admin
django-haystack==3.1.1
lxml
psycopg2
pysolr
regex
sqlparse
unicodecsv
pdfkit
